move Android version/display helpers from SeekerState to PlatformInfo
