+++ Seeker/Main/SeekerState.cs
@@ -121,32 +121,6 @@ namespace Seeker
 
         public static int NumberParsed = 0;
 
-        // TODO utils
-        public static bool RequiresEitherOpenDocumentTreeOrManageAllFiles()
-        {
-            //29 does has the requestExternalStorage workaround.
-            return OperatingSystem.IsAndroidVersionAtLeast(30);
-        }
-
-        public static bool UseLegacyStorage()
-        {
-            return !OperatingSystem.IsAndroidVersionAtLeast(29);
-        }
-
-
-        public static bool PreMoveDocument()
-        {
-            return !OperatingSystem.IsAndroidVersionAtLeast(24);
-        }
-
-        public static bool IsLowDpi()
-        {
-            return Android.Content.Res.Resources.System.DisplayMetrics.WidthPixels < 768;
-        }
-        // TODO utils
-
-
-
         // TODO hack?
         public static ManualResetEvent ManualResetEvent = new ManualResetEvent(false); //previously this was on the loginfragment but
                                                                                        //it would get recreated every time so there were lost instances with threads waiting forever....
