+++ Seeker/Transfers/TransfersFragment.cs
@@ -813,7 +813,7 @@ namespace Seeker
                             //tested on API25 and API30
                             //AndroidX.Core.Content.FileProvider
                             Android.Net.Uri uriToUse = null;
-                            if (SeekerState.UseLegacyStorage() && SimpleHelpers.IsFileUri((tItem as TransferItem).FinalUri)) //i.e. if it is a FILE URI.
+                            if (PlatformInfo.UseLegacyStorage() && SimpleHelpers.IsFileUri((tItem as TransferItem).FinalUri)) //i.e. if it is a FILE URI.
                             {
                                 uriToUse = AndroidX.Core.Content.FileProvider.GetUriForFile(this.Context, this.Context.ApplicationContext.PackageName + ".provider", new Java.IO.File(Android.Net.Uri.Parse((tItem as TransferItem).FinalUri).Path));
                             }
